Use node:test for testing
- node:test に移行し、ルール数・キー名一致・lint 実行の3テストを追加
- テストファイルを preset.test.mjs にリネーム（node --test の規約に準拠）
